# combi-uees-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "COMBI UEES con Laravel 12",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/anuncios/{anuncio}": {
            "put": {
                "tags": [
                    "Anuncios"
                ],
                "summary": "Actualizar la URL de la foto del anuncio",
                "operationId": "93c86ed4353d8363bca3f88d99ee97c4",
                "parameters": [
                    {
                        "name": "anuncio",
                        "in": "path",
                        "description": "ID del anuncio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "anuncioURLFoto"
                                ],
                                "properties": {
                                    "anuncioURLFoto": {
                                        "type": "string",
                                        "format": "url"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Anuncio actualizado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "operationId": "80c0b886319fc75046ae8c50ca42221e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "remember": {
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "a90ec8a03ca670dc930ffa583f08b202",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada exitosamente"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/anuncios/{anuncio}/comentarios": {
            "get": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Listar comentarios de un anuncio",
                "operationId": "470792416a2bf9ac55c1e0ceb40ce107",
                "parameters": [
                    {
                        "name": "anuncio",
                        "in": "path",
                        "description": "ID del anuncio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de comentarios"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Crear un comentario en un anuncio",
                "operationId": "f91129b8389ffc930343fd8370b64ebc",
                "parameters": [
                    {
                        "name": "anuncio",
                        "in": "path",
                        "description": "ID del anuncio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombreUsuario",
                                    "comentarioTitulo",
                                    "comentarioDescripción"
                                ],
                                "properties": {
                                    "nombreUsuario": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "comentarioTitulo": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "comentarioDescripción": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comentario registrado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/rutas/{ruta}/horarios": {
            "get": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Obtener todos los horarios de una ruta específica",
                "operationId": "445e42400ad0066037deada60d9fdac4",
                "parameters": [
                    {
                        "name": "ruta",
                        "in": "path",
                        "description": "ID de la ruta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de horarios en orden descendente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Horario"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Crear un nuevo horario para una ruta",
                "operationId": "dd30a4d5480ec185b0d0c0a081399b4d",
                "parameters": [
                    {
                        "name": "ruta",
                        "in": "path",
                        "description": "ID de la ruta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "horaSalida"
                                ],
                                "properties": {
                                    "horaSalida": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "14:30:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Horario creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Horario"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/horarios/{id}": {
            "delete": {
                "tags": [
                    "Horarios"
                ],
                "summary": "Eliminar un horario por su ID",
                "operationId": "427af64120bcdc63ed05235daf6731df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del horario a eliminar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Horario eliminado exitosamente"
                    },
                    "404": {
                        "description": "Horario no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Horario": {
                "title": "Horario",
                "required": [
                    "id",
                    "inicio",
                    "fin"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "inicio": {
                        "type": "string",
                        "format": "time"
                    },
                    "fin": {
                        "type": "string",
                        "format": "time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}